Split pattern struct
